pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analyzer"
authors = ["slavanr45 <slava.r45@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
repository = "https://github.com/slavanr45/python-project-83"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
